+++ src/WPGet_Elementor_Widgets/Application.php
@@ -19,14 +19,14 @@ class Application
     {
         add_action('elementor/editor/after_enqueue_scripts', [ $this, 'editor_enqueue_scripts'] );
         add_action('wp_enqueue_scripts', [$this, 'wp_enqueue_scripts']);
-
         add_action('elementor/elements/categories_registered', [$this, 'register_categories']);
-
-        /*@todo check for dependencies first */
 
         Module_Loader::instance();
     }
 
+    /*
+     * Widget Category for all of our Widgets
+     */
     public function register_categories( Elements_Manager $elements_manager )
     {
         $elements_manager->add_category(
@@ -35,17 +35,18 @@ class Application
                 'title' => __('WPGet', WPG_WIDGETS_TEXT_DOMAIN),
                 'icon' => 'fa fa-plug'
             ]
-
         );
     }
 
     public function editor_enqueue_scripts()
     {
+        /* CSS jsu fro the Elementor Editor */
         wp_enqueue_style('wpg-style-editor', WPG_WIDGETS_URL . 'assets/css/editor.css');
     }
 
     public function wp_enqueue_scripts()
     {
+        /* Single CSS file for all Widgets, Alternately, you can add individual CSS files as Widget Dependencies */
         wp_enqueue_style('wpg-style-widgets', WPG_WIDGETS_URL . 'assets/css/ui-default.css');
     }