+++ src/WPGet_Elementor_Widgets/Module_Loader.php
@@ -2,6 +2,9 @@
 
 namespace WPGet_Elementor_Widgets;
 
+use WPGet_Elementor_Widgets\Helper\ElementorHelper;
+use WPGet_Elementor_Widgets\Lib\Module_Controller_Base;
+
 class Module_Loader
 {
     private static $_instance = null;
@@ -18,6 +21,7 @@ class Module_Loader
         $this->load_modules();
     }
 
+
     public function load_modules()
     {
         $sub_dirs = scandir(WPG_WIDGETS_MODULE_DIR);
@@ -25,15 +29,26 @@ class Module_Loader
         foreach ($sub_dirs as $dir){
             if(in_array($dir, ['.','..'])) continue;
 
-            //@todo add logic to load only if enabled
             $class_name =  __NAMESPACE__ . '\\Modules\\' . $dir . '\\'. 'Module_Controller';
 
-            if(class_exists($class_name)){
-                $class_name::instance();
-            }else{
-                // report warning here.
-            }
 
+            if(class_exists($class_name)){
+                /**@var $module Module_Controller_Base */
+                $module = $class_name::instance();
+                $conf = $module->get_config();
+                /*@todo add check for enable once admin panel has been done */
+                if(!empty($conf['plugin_dependencies'])){
+                    $ok = ElementorHelper::has_all_dependancies($conf['plugin_dependencies']);
+                    if($ok){
+                        $module->init();
+                    }else{
+                        //@todo report warning if dependencies aren't met, or just don't init ?
+                    }
+                }
+
+            }else{
+                //@todo report warning here if the class doesn't exist?.
+            }
         }
 
     }