+++ src/WPGet_Elementor_Widgets/Modules/Page_Settings_General/Module_Controller.php
@@ -4,9 +4,10 @@ namespace WPGet_Elementor_Widgets\Modules\Page_Settings_General;
 
 use Elementor\Plugin;
 use WPGet_Elementor_Widgets\Helper\ElementorHelper;
+use WPGet_Elementor_Widgets\Lib\Module_Controller_Base;
 
 
-class Module_Controller
+class Module_Controller extends Module_Controller_Base
 {
     private static $_instance = null;
 
@@ -19,9 +20,32 @@ class Module_Controller
 
     public function __construct()
     {
+        parent::__construct();
         Page_Settings::instance();
         add_filter('body_class', [$this, 'body_class']);
         add_action('wp_enqueue_scripts', [$this, 'load_scripts'],99);
+    }
+
+    /*
+     * Widgets registered here will register when init() is called
+     */
+    public function get_widgets()
+    {
+        return [];
+    }
+
+    /*
+     * Used by Module_Loader to check for plugin dependencies
+     * Also, will be used for info on the Admin panel
+     */
+    public function get_config()
+    {
+        return [
+            'plugin_dependencies' => ['elementor/elementor.php'],
+            'name' => 'Common Page Settings',
+            'description' => 'WPGet settings for the Page',
+            'can_disable' => false
+        ];
     }
 
     public function body_class($classes)